+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 +++++++++++++++++++
 
 
+- **Add** ``Script.get_context`` to get information where you currently are.
 - Big **Script API Changes**:
     - The line and column parameters of ``jedi.Script`` are now deprecated
     - ``completions`` deprecated, use ``complete`` instead